# contracts/openapi/zen70-gateway-full.openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "ZEN70 API",
    "version": "1.58.0"
  },
  "paths": {
...
    "/api/v1/auth/invites": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Create Invite",
        "description": "生成一次性邀请凭证（仅管理员可用）",
        "operationId": "create_invite_api_v1_auth_invites_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InviteCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InviteResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/auth/invites/{token}/webauthn/register/begin": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Invite Webauthn Register Begin",
        "description": "带外传递链接 - 开始注册 WebAuthn",
        "operationId": "invite_webauthn_register_begin_api_v1_auth_invites__token__webauthn_register_begin_post",
        "parameters": [
          {
            "name": "token",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WebAuthnRegisterBeginResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/invites/{token}/webauthn/register/complete": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Invite Webauthn Register Complete",
        "description": "带外传递链接 - 完成 WebAuthn 注册并销毁 Token",
        "operationId": "invite_webauthn_register_complete_api_v1_auth_invites__token__webauthn_register_complete_post",
        "parameters": [
          {
            "name": "token",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Token"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WebAuthnRegisterCompleteRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Invite Webauthn Register Complete Api V1 Auth Invites  Token  Webauthn Register Complete Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/invites/{token}/fallback/login": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Invite Fallback Login",
        "description": "带外传递链接 - 大陆安卓设备降级免密直连并销毁 Token",
        "operationId": "invite_fallback_login_api_v1_auth_invites__token__fallback_login_post",
        "parameters": [
          {
            "name": "token",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Token"
            }
          },
          {
            "name": "X-Invite-Fallback-Confirm",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Invite-Fallback-Confirm"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Invite Fallback Login Api V1 Auth Invites  Token  Fallback Login Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "InviteCreateRequest": {
        "properties": {
          "user_id": {
            "type": "integer",
            "title": "User Id",
            "description": "目标用户ID"
          },
          "expires_in_minutes": {
            "type": "integer",
            "maximum": 1440.0,
            "minimum": 1.0,
            "title": "Expires In Minutes",
            "description": "有效时长(分钟)，1~1440(24小时)",
            "default": 15
          }
        },
        "type": "object",
        "required": [
          "user_id"
        ],
        "title": "InviteCreateRequest",
        "description": "管理员生成一次性邀请链接请求。"
      },
      "InviteResponse": {
        "properties": {
          "token": {
            "type": "string",
            "title": "Token",
            "description": "物理防腐的一次性 Token"
          },
          "expires_at": {
            "type": "integer",
            "title": "Expires At",
            "description": "过期时间戳"
          }
        },
        "type": "object",
        "required": [
          "token",
          "expires_at"
        ],
        "title": "InviteResponse",
        "description": "一次性邀请链接响应。"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
      "WebAuthnRegisterBeginResponse": {
        "properties": {
          "options": {
            "additionalProperties": true,
            "type": "object",
            "title": "Options",
            "description": "JSON 格式的 PublicKeyCredentialCreationOptions"
          }
        },
        "type": "object",
        "required": [
          "options"
        ],
        "title": "WebAuthnRegisterBeginResponse",
        "description": "WebAuthn 注册开始响应：options 可直接传给 navigator.credentials.create()。"
      },
      "WebAuthnRegisterCompleteRequest": {
        "properties": {
          "credential": {
            "additionalProperties": true,
            "type": "object",
            "title": "Credential",
            "description": "navigator.credentials.create() 的返回值"
          }
        },
        "type": "object",
        "required": [
          "credential"
        ],
        "title": "WebAuthnRegisterCompleteRequest",
        "description": "WebAuthn 注册完成请求：前端传来的 credential 对象。"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}